+++ src/com/genir/aitweaks/core/shipai/VentModule.kt
@@ -369,7 +369,7 @@ class VentModule(private val ai: CustomShipAI) {
 
     private fun waitForBursts(): Boolean {
         val burstWeapons = ship.allGroupedWeapons.filter { it.isBurstWeapon && !it.spec.isInterruptibleBurst }
-        val longestBursts: Float = burstWeapons.filter { it.isInBurst }.maxOfOrNull { it.burstFireTimeRemaining } ?: 0f
+        val longestBursts: Float = burstWeapons.filter { it.IsInBurst }.maxOfOrNull { it.burstFireTimeRemaining } ?: 0f
 
         // Vent when all the bursts have ended.
         if (longestBursts == 0f) {